Enhanced error handling for both clients.

+++ Casper.Network.SDK.Clients/ContractException.cs
@@ -4,14 +4,14 @@ namespace Casper.Network.SDK.Clients
 {
     public class ContractException : Exception
     {
-        public int Code { get; }
+        public long Code { get; }
         
-        public ContractException(string message, int code) : base(message)
+        public ContractException(string message, long code) : base(message)
         {
             Code = code;
         }
 
-        public ContractException(string message, int code, Exception innerException) : base(message, innerException)
+        public ContractException(string message, long code, Exception innerException) : base(message, innerException)
         {
             Code = code;
         }
